Add ability to assign/unassign within the Add Remove Members and Change Membership dialogs using the space bar. User may highlight a selection and click the space bar to assign/unasssign the selected objects. This makes these screens easier to use with the keyboard. Updated help documents to reflect this new ability.
Allow user to use CTRL-1, 2  or 3 to change between the tabs.

Changed default error message for when the action handler recieves an invalid/unsupported action.

Removed spurious < charcters from the Summary Report.

Modified Summary Report behavior so that it doesn't throw exception if a project is missing one or two project folders (branches, tags or trunk).